+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 2.3.5
+- Make Soul Fruit Blindness Effect configurable
 
 # 2.3.4
 - Give all BWG Entities Eye Heights